chapter 5. - добавлен HUD - вынесены настройки геймплея отдельно - добавлен подсчет пуль при стрельбе - добавлена музыка

+++ chapter5/core/src/com/github/br/paper/airplane/ecs/component/HealthComponent.java
@@ -4,10 +4,10 @@ import com.badlogic.ashley.core.Component;
 
 public class HealthComponent implements Component {
 
-    public int health;
-    public int damage;
+    public short health;
+    public short damage;
 
-    public HealthComponent(int health, int damage) {
+    public HealthComponent(short health, short damage) {
         this.health = health;
         this.damage = damage;
     }
